improve validation on delete about course instructor and student

+++ src/main/java/com/example/_2_ong_endy_pvh_spring_homework002/service/Impl/StudentServiceImpl.java
@@ -55,6 +55,9 @@ public class StudentServiceImpl implements StudentService {
 
     @Override
     public Student deleteStudentById(Long studentId) {
+        if (studentRepository.getStudentById(studentId) == null) {
+            return null;
+        }
         studentCourseRepository.deleteStudentCourseByStudentID(studentId);
         return studentRepository.deleteStudentById(studentId);
     }
